cpu: ip: move function definitions to cpp
